fix: update v0.2.0

+++ track003_java/bankproject/src/com/the703/v1/UserServiceImpl.java
@@ -10,28 +10,28 @@ public class UserServiceImpl implements UserService {
 	@Override
 	public boolean isExists(User user) {
 		User existUser = userRepo.selectByUserId(user.getUserId());
-		if(existUser != null && existUser.getPassword().equals(user.getPassword())) {
-			return true;
-		} else {
+		if(existUser == null || !existUser.getPassword().equals(user.getPassword())) {
 			System.out.println("=== 아이디 및 비밀번호를 다시 확인해주세요.");
 			return false;
 		}
+		return true;
 	}
 
 	@Override
 	public boolean isEmpty(User user) {
-		String userId = user.getUserId();
-		User existUser = userRepo.selectByUserId(userId);
-		if(user != null && existUser.getAccount() > user.getAccount()) {
-			return true;
-		} else {
+		User existUser = userRepo.selectByUserId(user.getUserId());
+		if(user == null || existUser.getAccount() < user.getAccount()) {
 			System.out.println("=== 계좌에 돈이 부족합니다.");
 			return false;
 		}
+		return true;
 	}
 
 	@Override
-	public User showUserInfo(String userId) {
+	public User showUserInfo(User user) {
+		if(!isExists(user)) return null;
+		
+		String userId = user.getUserId();
 		User existUser = userRepo.selectByUserId(userId);
 		
 		System.out.println("=== 계좌 조회");
@@ -52,6 +52,8 @@ public class UserServiceImpl implements UserService {
 
 	@Override
 	public int delUser(User user) {
+		if(!isExists(user)) return 0;
+
 		User delUser = userRepo.deleteByUserId(user.getUserId());
 		System.out.println("=== 계좌가 삭제되었습니다.");
 		
@@ -60,6 +62,8 @@ public class UserServiceImpl implements UserService {
 	
 	@Override
 	public int depositAccountByUserId(User user) {
+		if(!isExists(user)) return 0;
+		
 		User existUser = userRepo.updateAccountByUserId(user);
 		System.out.println("=== 입금 완료");
 		System.out.println("잔액: "+ existUser.getAccount());
@@ -69,8 +73,9 @@ public class UserServiceImpl implements UserService {
 	
 	@Override
 	public int withdrawalAccountByUserId(User user) {
-		user.setAccount(user.getAccount() * (-1));
-		
+		if(!isExists(user)) return 0;
+		if(!isEmpty(user)) return 0;
+
 		User existUser = userRepo.updateAccountByUserId(user);
 		System.out.println("=== 출금 완료");
 		System.out.println("잔액: "+ existUser.getAccount());
